Feat : DetailEdit 및  DetailDiv 통합
