+++ src/components/detail/DetailDiv.tsx
@@ -1,109 +1,154 @@
+import { useEffect, useState } from 'react';
 import { ProjectProps } from 'ProjectPageModule';
 import styled from 'styled-components';
 import MDEditor from '@uiw/react-md-editor';
-const DetailDiv: React.FunctionComponent<ProjectProps.IProjectProps> = (props) => {
-    const handleWindowOpen = (url: string) => {
-        window.open(url, '_blank');
+import { useParams } from 'react-router-dom';
+import { ImgEdit, TitleEdit, DateEdit, LinkEdit, TagEdit } from '../../components/edit';
+interface IProps {
+    handleChangeToggle?: (...args: any[]) => any;
+    editMode?: boolean;
+}
+
+const DetailEdit: React.FunctionComponent<IProps> = ({ handleChangeToggle, editMode }) => {
+    const projectId = useParams().id;
+    const data: ProjectProps.IProjectProps = {
+        projectId: projectId,
+        title: '프로젝트 토끼토끼',
+        startDate: '2021-01-31',
+        endDate: '2021-02-28',
+        techStack: ['Django', 'Flask', 'TypeScript'],
+        imgSrc: 'https://t1.daumcdn.net/cfile/tistory/996B5C3F5C2DCE5304?original',
+        gifSrc: 'https://t1.daumcdn.net/cfile/tistory/995040355C2DCE5E2E?original',
+        explain: '프로젝트설명0',
+        urlLink: [
+            { linkName: 'Live Demo', linkUrl: 'https://youtube.com' },
+            { linkName: 'Github', linkUrl: 'https://github.com' },
+        ],
     };
 
-    return (
-        <DetailContainer>
-            <TitleDiv>{props.title}</TitleDiv>
-            <p>{`제작 기간 : ${props.startDate.replace(/-/gi, '.')} ~ ${props.endDate.replace(/-/gi, '.')}`}</p>
-            <GifDiv>
-                <img src={`${props.gifSrc}`} alt="라이브 데모" />
-            </GifDiv>
-            <ExplainDiv>
-                <h2>프로젝트 설명</h2>
-                <MDEditor.Markdown source={props.explain} />
-            </ExplainDiv>
-            <StackDiv>
-                <h2>기술 스택</h2>
-                <TagDiv>
-                    {props.techStack.map((stack) => (
-                        <Tag>{stack}</Tag>
-                    ))}
-                </TagDiv>
-            </StackDiv>
+    const [title, setTitle] = useState<string>(data.title);
+    const [startDate, setStartDate] = useState<string>(data.startDate);
+    const [endDate, setEndDate] = useState<string>(data.endDate);
+    const [explain, setExplain] = useState<string>(data.explain);
+    const [gifSrc, setGifSrc] = useState<string>(data.gifSrc);
+    const [imgSrc, setImgSrc] = useState<string>(data.imgSrc);
+    const [techStack, setTechStack] = useState<string[]>(data.techStack);
+    const [urlLink, setUrlLink] = useState<ProjectProps.IUrl[]>(data.urlLink);
 
-            <LinkDiv>
-                {props.urlLink &&
-                    props.urlLink.map((url) => (
-                        <button onClick={() => handleWindowOpen(url.linkUrl ?? '')}>{url.linkName}</button>
-                    ))}
-            </LinkDiv>
-        </DetailContainer>
+    const handleSubmit = (event: any) => {
+        event.preventDefault();
+        const data: ProjectProps.IProjectProps = {
+            projectId: projectId,
+            title: title,
+            startDate: startDate,
+            endDate: endDate,
+            /*
+            gifSrc 와 imgSrc : 이미지 파일이 있으면 이미지 파일 Blob처리해서 보내고, 
+            건들지 않아서 서버에 이미지가 저장된 경로 그대로 보유하고 있다면, 그 경로만 보내준다.
+            */
+            gifSrc: gifBlob ?? gifSrc,
+            imgSrc: imgBlob ?? imgSrc,
+            techStack: techStack,
+            urlLink: urlLink,
+            explain: explain,
+        };
+        console.log(data);
+    };
+
+    const [gifBlob, setGifBlob] = useState<Blob>();
+    const [imgBlob, setImgBlob] = useState<Blob>();
+    const handleShowGifPreview = (event: any) => {
+        setGifSrc(URL.createObjectURL(event.target.files[0]));
+        setGifBlob(event.target.files[0]);
+    };
+    const handleShowImgPreview = (event: any) => {
+        setImgSrc(URL.createObjectURL(event.target.files[0]));
+        setImgBlob(event.target.files[0]);
+    };
+    const handleDeletePreview = (imgSrc: any, setImgSrc: Function) => {
+        URL.revokeObjectURL(imgSrc);
+        setImgSrc('');
+    };
+
+    useEffect(() => {
+        return () => {
+            handleDeletePreview(gifSrc, setGifSrc);
+            handleDeletePreview(imgSrc, setImgSrc);
+        };
+        //eslint-disable-next-line
+    }, []);
+
+    return (
+        <DetailForm onSubmit={handleSubmit}>
+            <TitleEdit title={title} setTitle={setTitle} editMode={editMode} />
+            <DateEdit
+                startDate={startDate}
+                endDate={endDate}
+                setStartDate={setStartDate}
+                setEndDate={setEndDate}
+                editMode={editMode}
+            />
+            <ImgEdit
+                editMode={editMode}
+                source={gifSrc}
+                alt="라이브 데모 GIF"
+                setSrc={setGifSrc}
+                handleShowPreview={handleShowGifPreview}
+                accept="gif,"
+            />
+            {editMode && (
+                <ImgEdit
+                    editMode={editMode}
+                    source={imgSrc}
+                    alt="프로젝트 이미지"
+                    setSrc={setImgSrc}
+                    handleShowPreview={handleShowImgPreview}
+                    accept="png, .jpg, .jpeg, .svg"
+                />
+            )}
+            <h2>프로젝트 설명</h2>
+            {editMode && (
+                <MDEditor
+                    height={window.innerHeight * 0.3}
+                    value={explain}
+                    onChange={(newValue = '') => setExplain(newValue)}
+                />
+            )}
+            <MDEditor.Markdown source={explain} />
+            <TagEdit techStack={techStack} setTechStack={setTechStack} editMode={editMode} />
+            <LinkEdit urlLink={urlLink} setUrlLink={setUrlLink} editMode={editMode} />
+            {editMode && (
+                <div>
+                    <button type="submit">수정</button>
+                    <button type="button" onClick={handleChangeToggle}>
+                        취소
+                    </button>
+                </div>
+            )}
+        </DetailForm>
     );
 };
 
-export default DetailDiv;
+export default DetailEdit;
 
-const DetailContainer = styled.div`
-    display: flex;
-    font-family: 'EliceRegular', 'Montserrat', 'Helvetica', 'sans-serif';
-    flex-direction: column;
-    padding: 0 5%;
-    text-align: left;
-    justify-content: center;
+const DetailForm = styled.form`
+    font-family: 'Montserrat', 'EliceRegular';
     margin-top: 3%;
-    p {
-        font-size: 1.5em;
-        color: #757575;
+    padding: 0 5%;
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+    label {
+        font-weight: bold;
     }
+
+    button {
+        border: 1px solid black;
+        margin: auto;
+    }
+
     h2 {
         border-bottom: 1px solid black;
-        margin-bottom: 2%;
-    }
-`;
-
-const TitleDiv = styled.div`
-    font-weight: bold;
-    font-size: 2em;
-`;
-const GifDiv = styled.div`
-    margin-top: 3%;
-    margin-bottom: 5%;
-    background-color: ${(props) => props.theme.color.background};
-    border-radius: 10px;
-    overflow: hidden;
-
-    img {
-        width: 100%;
-        object-fit: contain;
-    }
-`;
-const ExplainDiv = styled.div`
-    p {
-        color: #3a3a3a;
-    }
-`;
-
-const StackDiv = styled.div`
-    display: flex;
-    flex-direction: column;
-    margin-top: 5%;
-`;
-const TagDiv = styled.div`
-    display: flex;
-`;
-const Tag = styled.div`
-    margin: 2%;
-    background-color: ${(props) => props.theme.color.main};
-    color: ${(props) => props.theme.color.sub};
-    padding: 2%;
-`;
-
-const LinkDiv = styled.div`
-    justify-content: start;
-    font-family: 'Montserrat';
-    button {
-        margin: 5% 0;
-        margin-right: 5%;
-        width: 10em;
-        background-color: ${(props) => props.theme.color.background};
-        border-radius: 15px;
-        border: none;
-        font-weight: bold;
-        height: 3em;
+        margin: 2% 0;
     }
 `;
