+++ src/components/detail/index.ts
@@ -1,2 +1,2 @@
 export { default as DetailDiv } from './DetailDiv';
-export { default as DetailEdit } from './DetailEdit';
+export { default as ProjectAdd } from '../detail/ProjectAdd';
