+++ src/components/edit/TagEdit.tsx
@@ -1,12 +1,33 @@
+import styled from 'styled-components';
 interface Props {
-    closeModal: void;
+    techStack: string[];
+    setTechStack: (args: string[]) => void;
+    editMode?: boolean;
 }
-const LinkEdit = ({ closeModal }: Props) => {
+const LinkEdit = ({ techStack, setTechStack, editMode }: Props) => {
     return (
-        <div>
-            <button onClick={() => closeModal} />
-        </div>
+        <StackDiv>
+            <h2>기술 스택</h2>
+            <TagDiv>
+                {techStack && techStack.map((stack) => <Tag>{stack}</Tag>)}
+                {editMode && <Tag onClick={() => console.log('스택 추가')}>+</Tag>}
+            </TagDiv>
+        </StackDiv>
     );
 };
 
 export default LinkEdit;
+const StackDiv = styled.div`
+    display: flex;
+    flex-direction: column;
+    margin-top: 5%;
+`;
+const TagDiv = styled.div`
+    display: flex;
+`;
+const Tag = styled.div`
+    margin: 2%;
+    background-color: ${(props) => props.theme.color.main};
+    color: ${(props) => props.theme.color.sub};
+    padding: 2%;
+`;
