+++ src/components/project/index.ts
@@ -1,4 +1,3 @@
 export { default as ProjectCard } from './ProjectCard';
 export { default as ProjectDiv } from './ProjectDiv';
 export { default as ProjectEdit } from './ProjectEdit';
-export { default as ProjectAdd } from './ProjectAdd';
